% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurchaserConfirm.R
\name{PurchaserConfirmUpdateServer}
\alias{PurchaserConfirmUpdateServer}
\title{处理逻辑}
\usage{
PurchaserConfirmUpdateServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
处理逻辑
}
\examples{
PurchaserConfirmUpdateServer()
}
